+++ charts/datadog/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Datadog changelog
 
+## 3.223.3
+
+* Bump the default App & API Protection (AppSec) sidecar processor image tag (`datadog.appsec.injector.sidecar.imageTag`) from `v2.6.0` to `v2.8.2`.
+
+## 3.223.2
+
+* Clarify `datadog.env` and `datadog.envDict` that environment variables set with these options only propagate to the node Agent containers only.
+
+## 3.223.1
+
+* Update Default Agent Version to 7.80.1.
+
 ## 3.223.0
 
 * [CONTP-1710] Install DatadogInstrumentation CRD and add RBAC permissions when controller is enabled ([#2717](https://github.com/DataDog/helm-charts/pull/2717)).